Fix freebsd lint
