+++ cmd/nerdctl/container/container_cp_linux.go
@@ -17,7 +17,10 @@
 package container
 
 import (
+	"errors"
 	"fmt"
+	"path/filepath"
+	"strings"
 
 	"github.com/spf13/cobra"
 
@@ -130,4 +133,47 @@ func processCpOptions(cmd *cobra.Command, args []string) (types.ContainerCpOptio
 
 func AddCpCommand(rootCmd *cobra.Command) {
 	rootCmd.AddCommand(newCpCommand())
+}
+
+var errFileSpecDoesntMatchFormat = errors.New("filespec must match the canonical format: [container:]file/path")
+
+func parseCpFileSpec(arg string) (*cpFileSpec, error) {
+	i := strings.Index(arg, ":")
+
+	// filespec starting with a semicolon is invalid
+	if i == 0 {
+		return nil, errFileSpecDoesntMatchFormat
+	}
+
+	if filepath.IsAbs(arg) {
+		// Explicit local absolute path, e.g., `C:\foo` or `/foo`.
+		return &cpFileSpec{
+			Container: nil,
+			Path:      arg,
+		}, nil
+	}
+
+	parts := strings.SplitN(arg, ":", 2)
+
+	if len(parts) == 1 || strings.HasPrefix(parts[0], ".") {
+		// Either there's no `:` in the arg
+		// OR it's an explicit local relative path like `./file:name.txt`.
+		return &cpFileSpec{
+			Path: arg,
+		}, nil
+	}
+
+	return &cpFileSpec{
+		Container: &parts[0],
+		Path:      parts[1],
+	}, nil
+}
+
+type cpFileSpec struct {
+	Container *string
+	Path      string
+}
+
+func cpShellComplete(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	return nil, cobra.ShellCompDirectiveFilterFileExt
 }
